refactor: restructure application view and enhance organization query handling
