+++ web/src/lib/hooks/use-organization.tsx
@@ -9,9 +9,14 @@ import type {
 import { useApiQuery } from '@/lib/hooks/use-api';
 import { platformApiPath } from '@/lib/platform-api';
 import { useUserProfile } from '@/lib/hooks/use-user';
-import { ApiError, apiQueryKey, fetchApiJson, requestApi, requestApiJson } from '@/lib/api';
+import { ApiError, fetchApiJson, requestApi, requestApiJson } from '@/lib/api';
 import { zOrganizationDetails, zOrganizationSummary } from '@/lib/generated/platform-api-v1/zod.gen';
-import { applicationsQueryKey, organizationsQueryKey, userOrganizationsQueryKey } from '@/lib/query-keys';
+import {
+    applicationsQueryKey,
+    organizationQueryKey,
+    organizationsQueryKey,
+    userOrganizationsQueryKey,
+} from '@/lib/query-keys';
 
 /** Fetches organization details and related collections for the current workspace. */
 export function useOrganization(organizationSlug: string) {
@@ -49,7 +54,6 @@ export function useOrganization(organizationSlug: string) {
 /** Provides mutations for organization members and invitations. */
 export function useOrganizationMembers(organizationId: string) {
     const queryClient = useQueryClient();
-    const organizationPath = platformApiPath(`/organizations/${organizationId}`);
 
     const inviteMember = useMutation({
         mutationFn: async (payload: OrganizationInvitationCreate) => {
@@ -62,7 +66,7 @@ export function useOrganizationMembers(organizationId: string) {
                 method: 'POST',
             });
         },
-        onSuccess: () => queryClient.invalidateQueries({ queryKey: apiQueryKey(organizationPath) }),
+        onSuccess: () => queryClient.invalidateQueries({ queryKey: organizationQueryKey(organizationId) }),
     });
 
     const changeMemberRole = useMutation({
@@ -83,7 +87,7 @@ export function useOrganizationMembers(organizationId: string) {
         onSuccess: () =>
             Promise.all([
                 queryClient.invalidateQueries({ queryKey: userOrganizationsQueryKey }),
-                queryClient.invalidateQueries({ queryKey: apiQueryKey(organizationPath) }),
+                queryClient.invalidateQueries({ queryKey: organizationQueryKey(organizationId) }),
             ]),
     });
 
@@ -93,7 +97,6 @@ export function useOrganizationMembers(organizationId: string) {
 /** Creates one application and refreshes organization application data. */
 export function useCreateOrganizationApplication(organizationId: string) {
     const queryClient = useQueryClient();
-    const organizationPath = platformApiPath(`/organizations/${organizationId}`);
 
     return useMutation({
         mutationFn: async (payload: ApplicationCreate) => {
@@ -108,7 +111,7 @@ export function useCreateOrganizationApplication(organizationId: string) {
         },
         onSuccess: () =>
             Promise.all([
-                queryClient.invalidateQueries({ queryKey: apiQueryKey(organizationPath) }),
+                queryClient.invalidateQueries({ queryKey: organizationQueryKey(organizationId) }),
                 queryClient.invalidateQueries({ queryKey: applicationsQueryKey }),
             ]),
     });
@@ -117,7 +120,6 @@ export function useCreateOrganizationApplication(organizationId: string) {
 /** Deletes one application and refreshes organization application data. */
 export function useDeleteOrganizationApplication(organizationId: string) {
     const queryClient = useQueryClient();
-    const organizationPath = platformApiPath(`/organizations/${organizationId}`);
 
     return useMutation({
         mutationFn: async (applicationId: string) => {
@@ -130,7 +132,7 @@ export function useDeleteOrganizationApplication(organizationId: string) {
         },
         onSuccess: () =>
             Promise.all([
-                queryClient.invalidateQueries({ queryKey: apiQueryKey(organizationPath) }),
+                queryClient.invalidateQueries({ queryKey: organizationQueryKey(organizationId) }),
                 queryClient.invalidateQueries({ queryKey: applicationsQueryKey }),
             ]),
     });
@@ -159,7 +161,7 @@ export function useUpdateOrganization(organizationId: string) {
         // Refresh every response that embeds Organization metadata.
         onSuccess: () =>
             Promise.all([
-                queryClient.invalidateQueries({ queryKey: apiQueryKey(organizationPath) }),
+                queryClient.invalidateQueries({ queryKey: organizationQueryKey(organizationId) }),
                 queryClient.invalidateQueries({ queryKey: applicationsQueryKey }),
                 queryClient.invalidateQueries({ queryKey: organizationsQueryKey }),
                 queryClient.invalidateQueries({ queryKey: userOrganizationsQueryKey }),
